requirements.txt
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
filelock==3.15.4
fsspec==2024.6.1
gdown==5.2.0
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy<2
pillow==10.4.0
PySocks==1.7.1
requests==2.32.3
soupsieve==2.5
sympy==1.12.1
torch
torchaudio
torchvision
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
mlflow
tensorboard
wandb
torchmetrics
torchmetrics[detection]